Add basic ENet networking: client/server, UI integration
Introduce foundational networking with ENet, including client and server classes, peer management, and integration into the game's state/UI. The Game class now manages networking objects. Host and Join states create and connect clients/servers as needed. Refactored ServerPeerlist to use PeerEntry for better peer tracking. Added connection/disconnection logic, event handling, and improved documentation. Lays groundwork for multiplayer support.

+++ EchoDungeon/Networking/NetworkConstants.h
@@ -9,7 +9,7 @@ namespace NetworkConstants {
 	constexpr const int MAX_SIMULTANEOUS_CONNECTIONS = 4;
 
 	// Default port for networking (Server/Client)
-	constexpr const uint16_t DEFAULT_PORT = 7422;
+	constexpr const int DEFAULT_PORT = 7422;
 
 	// Maximum number of channels for ENet communication (Server/Client)
 	constexpr const int MAX_CHANNELS = 2;
